Modified position and velocity algorithms.
Changed the way that position, velocity and acceleration are calculated and stored. The object now keeps track of its previous, current, and future position, velocity, and acceleration in nine variables. Calling the numerical integration methods beemanPosition(), beemanAcceleration(), and beemanVelocity() from the Simulation object (in this order only!) will calculate and store the next position, acceleration, and velocity from the existing position and the positions from other bodies in the system. Calling the stepForward() method will move the object to the newly calculated location with appropriate velocity and acceleration and modify the relevant variables so the that object's new current position can be plotted. The future position, velocity, and acceleration variables are not overwritten until the numerical integration methods are called again!